main/requirements.txt
discord.py==2.3.2
getmac==0.9.4
pycryptodome==3.18.0
py-cpuinfo==9.0.0
